+++ tools/automation/utilities/path.py
@@ -6,7 +6,7 @@
 import os
 import glob
 
-from automation.utilities.const import COMMAND_MODULE_PREFIX
+from automation.utilities.const import COMMAND_MODULE_PREFIX, EXTENSIONS_MOD_PREFIX
 
 
 def get_repo_root():
@@ -26,6 +26,21 @@ def get_all_module_paths():
 def get_config_dir():
     """ Returns the users Azure directory. """
     return os.getenv('AZURE_CONFIG_DIR', None) or os.path.expanduser(os.path.join('~', '.azure'))
+
+
+def get_extension_dir():
+    """ Returns the extensions directory. """
+    custom_dir = os.environ.get('AZURE_EXTENSION_DIR')
+    return os.path.expanduser(custom_dir) if custom_dir else os.path.join(get_config_dir(), 'cliextensions')
+
+
+def get_extensions_paths(include_prefix=False):
+    glob_pattern = os.path.normcase('/*/{}*'.format(EXTENSIONS_MOD_PREFIX))
+    for path in glob.glob(get_extension_dir() + glob_pattern):
+        name = os.path.basename(path)
+        if not include_prefix:
+            name = name[len(EXTENSIONS_MOD_PREFIX):]
+        yield name, path
 
 
 def get_command_modules_paths(include_prefix=False):