+++ src/command_modules/azure-cli-storage/HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ===============
 
+2.2.5
++++++
+* Improve handling of corner cases for storage copy commands.
+* Fix issue with `storage blob copy start-batch` not using login credentials when the destination and source accounts are the same.
+* `storage blob/file url`- fix bug with sas_token not being incorporated into url.
+* Warn users about future breaking change: `blob/container list` will output first 5000 results by default.
+
 2.2.4
 +++++
 * Allow connection to storage services only with SAS and endpoints (without an account name or a key) as described in